backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
pytest
psycopg2-binary
pymongo
python-jose
PyJWT
python-dotenv